made in Python language
1)Code is basically been made in Python language.
2)We have used TextBlob library to determine the parts of speech of each
word.
3)In this program, we have made three functions with the main section of
reading data from file character by character.
Such that all pronouns contained in between & of symbol are collected in
a list named as "pronoun".
All lines of passage are contained in a list named as "para".
All nouns which are to be matched are in list-'noun_list'.
4)Among the three functions...
manipulation function is to check if the pronoun in between & symbol is
pronoun and if it is a pronoun than we are sending in different pronoun
cases...
while in check_conjunction function we are checking all conjunction in a
line, if their is any conjunction than we are doing some manipulations..
In in_between function we are doing some manipulations related to
backtracking and many other.
5)As a result if the program is able to find noun for the given pronoun,
than it is displayed. Else it display "Noun not available".